[CL-2110] Fix styles
