+++ front/app/modules/commercial/id_cow/components/VerificationFormCOW.tsx
@@ -20,7 +20,7 @@ import {
   SubmitButton,
   CancelButton,
   HelpImage,
-} from 'components/SignUpIn/SignUp/VerificationSignUpStep/styles';
+} from 'components/AuthProviders/styles';
 
 // hooks
 import useAuthUser from 'hooks/useAuthUser';
