+++ front/app/modules/commercial/id_id_card_lookup/components/IdCardLookupForm.tsx
@@ -20,7 +20,7 @@ import {
   SubmitButton,
   CancelButton,
   HelpImage,
-} from 'components/SignUpIn/SignUp/VerificationSignUpStep/styles';
+} from 'components/AuthProviders/styles';
 
 // hooks
 import useAuthUser from 'hooks/useAuthUser';
