+++ front/app/modules/commercial/id_oostende_rrn/components/VerificationFormOostendeRrn.tsx
@@ -19,7 +19,7 @@ import {
   Footer,
   SubmitButton,
   CancelButton,
-} from 'components/SignUpIn/SignUp/VerificationSignUpStep/styles';
+} from 'components/AuthProviders/styles';
 
 // hooks
 import useAuthUser from 'hooks/useAuthUser';
